Extract common resolution logic into traits
- ResolvesFromInfo: getDefinitionLocation(), getDocumentation(), format()
- ResolvesCallableParameters: getParameters(), getReturnType(), getParameterAtPosition(), getParameterByName()
