+++ src/Resolution/ResolvedClass.php
@@ -6,30 +6,17 @@ namespace Firehed\PhpLsp\Resolution;
 
 use Firehed\PhpLsp\Domain\ClassInfo;
 use Firehed\PhpLsp\Domain\Type;
-use Firehed\PhpLsp\Index\Location;
-use Firehed\PhpLsp\Utility\DocblockParser;
 
 /**
  * A resolved class, interface, trait, or enum wrapping ClassInfo.
  */
 final readonly class ResolvedClass implements ResolvedSymbol
 {
+    use ResolvesFromInfo;
+
     public function __construct(
         private ClassInfo $info,
     ) {
-    }
-
-    public function getDefinitionLocation(): ?Location
-    {
-        return Location::fromFileLine($this->info->file, $this->info->line);
-    }
-
-    public function getDocumentation(): ?string
-    {
-        if ($this->info->docblock === null) {
-            return null;
-        }
-        return DocblockParser::extractDescription($this->info->docblock);
     }
 
     /**
@@ -39,10 +26,5 @@ final readonly class ResolvedClass implements ResolvedSymbol
     public function getType(): Type
     {
         return $this->info->name;
-    }
-
-    public function format(): string
-    {
-        return $this->info->format();
     }
 }
