+++ src/Resolution/ResolvedEnumCase.php
@@ -7,8 +7,6 @@ namespace Firehed\PhpLsp\Resolution;
 use Firehed\PhpLsp\Domain\ClassName;
 use Firehed\PhpLsp\Domain\EnumCaseInfo;
 use Firehed\PhpLsp\Domain\Type;
-use Firehed\PhpLsp\Index\Location;
-use Firehed\PhpLsp\Utility\DocblockParser;
 
 /**
  * A resolved enum case wrapping EnumCaseInfo.
@@ -19,22 +17,11 @@ use Firehed\PhpLsp\Utility\DocblockParser;
  */
 final readonly class ResolvedEnumCase implements ResolvedSymbol
 {
+    use ResolvesFromInfo;
+
     public function __construct(
         private EnumCaseInfo $info,
     ) {
-    }
-
-    public function getDefinitionLocation(): ?Location
-    {
-        return Location::fromFileLine($this->info->file, $this->info->line);
-    }
-
-    public function getDocumentation(): ?string
-    {
-        if ($this->info->docblock === null) {
-            return null;
-        }
-        return DocblockParser::extractDescription($this->info->docblock);
     }
 
     /**
@@ -43,11 +30,6 @@ final readonly class ResolvedEnumCase implements ResolvedSymbol
     public function getType(): Type
     {
         return $this->info->declaringClass;
-    }
-
-    public function format(): string
-    {
-        return $this->info->format();
     }
 
     /**
